Add: Best Effort shorthand serialization
